[#69]feat: 테스트폴더 구조 정리
